backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
dj-rest-auth[with_social]>=5.0.0
django-allauth>=0.54.0
pandas
openpyxl
